backend/requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.1.4
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==38.2.0
fonttools==4.60.1
Markdown==3.10
openpyxl==3.1.5
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-decouple==3.8
sqlparse==0.5.3
tinycss2==1.5.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
